pyproject.toml
[project]
name = "portfolio-management"
version = "0.1.0"
description = "Tax-optimized portfolio management platform"
readme = "README.md"
requires-python = ">=3.14"
authors = [
    { name = "Charles Beck III", email = "charlesbeck@gmail.com" }
]
dependencies = [
    "django>=6.0,<6.1",              # Django LTS - update quarterly
    "pandas>=2.3.3",                  # Data manipulation - update monthly
    "python-dotenv>=1.0,<2.0",       # Environment config - stable
    "structlog>=25.5.0",              # Structured logging - update as needed
    "whitenoise[brotli]>=6.11.0",    # Static files - update with Django
    "yfinance>=0.2.66",               # Market data - update monthly
    "psycopg[binary]>=3.2.0",         # PostgreSQL adapter (v3 for async support)
]

[tool.mypy]
python_version = "3.14"
plugins = ["mypy_django_plugin.main"]

# Strictness settings
check_untyped_defs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true

# Warnings
warn_redundant_casts = true
warn_unused_ignores = true
warn_unused_configs = true
warn_return_any = true
warn_unreachable = true

# Error handling
ignore_missing_imports = true
strict_optional = true
show_error_codes = true

# Disable type-arg warnings for Django classes that aren't subscriptable at runtime
# See: https://github.com/python/mypy/issues/4717
disable_error_code = ["type-arg"]

# Exclude patterns
exclude = [
    "portfolio/migrations/",
    "portfolio/tests/",
    "portfolio/management/commands/",
    "scripts/",
    "conftest.py",
]

[[tool.mypy.overrides]]
module = "portfolio.tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "portfolio.migrations.*"
ignore_errors = true

# Django-heavy modules: disable strict checks for Django ORM attributes
[[tool.mypy.overrides]]
module = [
    "portfolio.admin",
    "portfolio.forms.*",
    "portfolio.views.*",
    "portfolio.domain.*",
    "portfolio.services.seeder",
    "portfolio.services.target_allocations",
]
disallow_untyped_defs = false
disallow_untyped_calls = false
warn_return_any = false

# Services with Django ORM queries: allow .objects attribute
[[tool.mypy.overrides]]
module = "portfolio.services.allocation_calculations"
warn_return_any = false

[tool.django-stubs]
django_settings_module = "config.settings.base"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "playwright>=1.56.0",            # E2E testing - update quarterly
    "pytest-django>=4.11.1",         # Testing - update quarterly
    "pytest-cov>=6.0.0",             # Coverage plugin - new dependency
    "pytest-playwright>=0.7.2",      # E2E testing - update with Playwright
    "django-debug-toolbar>=4.4",     # Debug tools - update with Django
    "ruff>=0.7",                     # Linting - update monthly
    "mypy>=1.13",                    # Type checking - update monthly
    "django-stubs>=5.1",             # Type stubs - update with Django
    "pre-commit>=4.0",               # Git hooks - update quarterly
    "coverage>=7.0",                 # Test coverage - update as needed
    "detect-secrets>=1.4.0",         # Secret scanning - update quarterly
    "safety>=3.7.0",                 # Security scanning - update monthly
    "factory-boy>=3.3.3",
    "types-pytz>=2025.2.0.20251108",
]

[tool.hatch.build.targets.wheel]
packages = ["portfolio", "config"]

[tool.ruff]
# Line length
line-length = 100

# Python version
target-version = "py312"

# Exclude directories
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "migrations",
    "staticfiles",
    "media",
]

[tool.ruff.lint]
# Enable specific rule sets
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "DJ",  # flake8-django
    "SIM", # flake8-simplify
]

# Ignore specific rules
ignore = [
    "E501",  # Line too long (handled by formatter)
    "B008",  # Do not perform function calls in argument defaults (Django uses this pattern)
]

# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
# Tests can use magic values and assertions
"tests/*.py" = ["PLR2004", "S101"]
# Settings files can have wildcard imports
"*/settings.py" = ["F405"]
# Admin files can have long lines for list_display
"*/admin.py" = ["E501"]
# Ignore Bandit alerts in tests (asserts, hardcoded passwords, pseudo-randomness)
"tests/*" = ["S101", "S105", "S106", "S311", "PLR2004"]
"**/tests/*" = ["S101", "S105", "S106", "S311", "PLR2004"]
"**test_*.py" = ["S101", "S105", "S106", "S311", "PLR2004"]
"**/conftest.py" = ["S101", "S105", "S106", "S311", "PLR2004"]

[tool.ruff.lint.isort]
# Django-specific isort configuration
known-first-party = ["portfolio", "config"]
section-order = [
    "future",
    "standard-library",
    "django",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.isort.sections]
"django" = ["django"]

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Use spaces for indentation
indent-style = "space"

# Like Black, respect magic trailing commas
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending
line-ending = "auto"

[tool.django]
# Django-specific settings for tools that support it
settings-module = "config.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.testing"
python_files = ["test_*.py", "*_test.py"]
testpaths = ["portfolio/tests"]
addopts = "--cov=portfolio --cov-report=term-missing --cov-report=html:coverage_html"

[tool.coverage.run]
# Coverage configuration
source = ["."]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/venv/*",
    "*/.venv/*",
    "manage.py",
    "config/wsgi.py",
    "config/asgi.py",
]

[tool.coverage.report]
# Coverage reporting
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
